cleanup features and CI
