+++ Cargo.toml
@@ -20,7 +20,6 @@ members = [
 ]
 
 [features]
-default = ["serde"]
 llvm-intrinsics = ["ethnum-intrinsics"]
 
 [dependencies]
